start testing p2p data storage
