+++ build.gradle
@@ -40,7 +40,7 @@ configure(subprojects) {
         bitcoinjVersion = 'a88d36d'
         logbackVersion = '1.1.10'
         lombokVersion = '1.18.2'
-        mockitoVersion = '2.21.0'
+        mockitoVersion = '2.13.0'
         powermockVersion = '2.0.0-beta.5'
         protobufVersion = '3.5.1'
         slf4jVersion = '1.7.22'
@@ -215,6 +215,8 @@ configure(project(':p2p')) {
         testCompile 'org.jmockit:jmockit:1.30' // must not use current $jmockitVersion
         testCompileOnly "org.projectlombok:lombok:$lombokVersion"
         testAnnotationProcessor "org.projectlombok:lombok:$lombokVersion"
+        testCompile("org.mockito:mockito-core:$mockitoVersion") {
+        }
     }
 }
 
@@ -251,7 +253,6 @@ configure(project(':core')) {
 
         testCompile "org.jmockit:jmockit:$jmockitVersion"
         testCompile("org.mockito:mockito-core:$mockitoVersion") {
-            exclude(module: 'objenesis')
         }
         testCompile "org.powermock:powermock-module-junit4:$powermockVersion"
         testCompile "org.powermock:powermock-api-mockito2:$powermockVersion"
